[PLAT-870] Problem: stats for types of propfinds not distinct
Solution: Account for their number and latency distinctly.

With the Prometheus Statsd exporter in place, we will have real histograms that
are fed by `stats_timer`, which means we can ditch the calls to `stats_counter`
and we can just reconfigure the bucket widths in the exporter config rather than
having a const that says how long is too long.

Therefore, If the Prometheus Statsd thing pans out, we can eliminate all the
`stats_{counter,timer}` calls in this function that weren't added in this
change.