[PLAT-870] Problem: stats for types of requests not distinct
Solution: Account for their number and latency distinctly.

With the Prometheus Statsd exporter in place, we have real histograms that are
fed by `stats_timer`, which means we can ditch related calls to `stats_counter`
and we can just reconfigure the bucket widths in the exporter config rather than
having a const that says how long is too long.

We instrument in `print_errors` because it has the error and latency information
as well as the function name. To that end, I've changed the function name passed
to `process_status` to more closely align with what someone looking at logs or
timeseries might understand because

1) that's easier than more extensive refactoring
2) the other people who care are the ones reading the source and they can just
as easily search for calls to `print_errors`